define some framework/systemic interfaces that will be necessary. more needed.
The general idea is that state machines can be defined centrally and downloaded for local test.
Also, it would be nice if the various logic modules can expose/document their logic via state machines.

Also, it would be nice to have a mini-language with which we can define conditions
so that guard conditions could be exposed.

Currently considering gRpc and proto3 as the messaging serialisation and transport mechanism.